ml-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
opencv-python
numpy
ultralytics
transformers
accelerate
bitsandbytes
torch
torchvision
imagehash
exifread
qdrant-client
python-dotenv
pydantic
requests